병상배정 관련 테이블 이력코드,순번 pk에서 제거, not null 제거 NewsScoreParam 수정

+++ src/main/kotlin/org/sbas/dtos/InfoPtDto.kt
@@ -53,10 +53,10 @@ fun InfoPtDto.toEntity(): InfoPt {
 @NoArg
 data class NewsScoreParam(
     var breath: Int,  // 분당호흡수
-    var spo2: Int,    // 산소포화도
+    var spo2: Double,    // 산소포화도
     var o2Apply: String, // 산소투여여부
     var sbp: Int,     // 수축기혈압
     var pulse: Int,   // 맥박
     var avpu: String,    // 의식상태
-    var bdTemp: Float,  // 체온
+    var bdTemp: Double,  // 체온
 )
